==> 3_batch_removal_Combat.R <==
library("ChAMP")
library(ggplot2)
library(ggfortify)
MyDir <- "C:/Work/Anna_Plaksienko/Anna_methylation/Data/"
MyOutputDir <- paste0(MyDir,"ChAMP_Output/")

load(paste0(MyOutputDir, "mybetaNorm.Rdata"))

champ.SVD(beta = as.data.frame(mybetaNorm), pd = pd,
          resultsDir = paste0(MyOutputDir, "CHAMP_SVDimages/"))

#we correct only for run and array, since that's what we saw in PCA
#there is correlation with slide, but mostly because it is very strong for one slide (Run 2)
#although we can also see it for other slides too
mybetaCombat_run_array <- champ.runCombat(beta = mybetaNorm,
                                          pd = pd,
                                          variablename = "Sample_Group",
                                          batchname = c("Run", "Array"))

mybetaCombat_slide_array <- champ.runCombat(beta = mybetaNorm,
                                          pd = pd,
                                          variablename = "Sample_Group",
                                          batchname = c("Slide", "Array"))

champ.SVD(beta = as.data.frame(mybetaCombat_slide_array), pd = pd,
          resultsDir = paste0(MyOutputDir,"CHAMP_SVDimages_Combat/"))

#PCA plots to see the effect of Combat
{
  pca_cl <- prcomp(t(log2(mybetaCombat_slide_array)), scale. = TRUE)
  #summary(pca_cl)
  
  #was generated by SVD before
  #screeplot(pca_cl, type = "l", main = "PCA scree plot")
  
  PCA <- as.data.frame(pca_cl$x)
  PCA <- cbind(PCA, pd[, c("Slide", "Array", "Run",
                           "Tempo", "Stato", "Stimolo", 
                           "Sample_Group")])
  
  autoplot(pca_cl, data=PCA, colour = "Array", shape = "Run",
           size = 3) + theme(text = element_text(size = 14))
  
  autoplot(pca_cl, data = PCA, colour = "Slide",
           size = 3) +
    theme(text = element_text(size = 14))
  
}

save(list = c("mybetaCombat_run_array", "mybetaCombat_slide_array", "pd"),
     file = paste0(MyOutputDir, "mybetaCombat_run_slide_array.Rdata"))
